10 - Image Captioning/deployment/requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.5.1%2Bcpu-cp38-cp38-linux_x86_64.whl
numpy==1.18.5
torchvision==0.6.1
Pillow==7.2.0
requests_toolbelt